src/TrafficLight.java
public class TrafficLight {

    private int id;
    private int numberOfLights;
    private int defaultColor;

    public TrafficLight() {
        id = 0;
        numberOfLights = 3;
        defaultColor = 1;
    }

    public TrafficLight(int id, int numberOfLights) {
        this.id = id;
        this.numberOfLights = numberOfLights;
    }

    public String showColor(int light) {

        String result;
        switch (light) {
            case 1:
                System.out.println("Red");
                result = "Red";
                break;
            case 2:
                System.out.println("Orange");
                result = "Orange";
                break;
            case 3:
                System.out.println("Green");
                result = "Green";
                break;
            default:
                System.out.println("Red");
                result = "red";
                break;
        }
        return result;
    }

    public int showId() {
       return id;
    }

    public void showColorAndId(int light) {
        String result;
        result = showColor(light);

        System.out.println("My color is "
                + result
                + " and my id is: "
                + showId()
        );
    }

    public boolean isPositiveNumber(int number) {
        return number >= 0;
    }

}
